Core/PacketIO: Updated packet structures to 11.0.5
Port From (https://github.com/TrinityCore/TrinityCore/commit/5f37b29f41b6cb3fbcc3dc6bc0cf3a25c1f7eb6a)

+++ Source/Game/Networking/Packets/ItemPackets.cs
@@ -76,7 +76,7 @@ namespace Game.Networking.Packets
         {
             _worldPacket.WritePackedGuid(VendorGUID);
             _worldPacket.WriteUInt32(Muid);
-            _worldPacket.WriteUInt8((byte)Reason);
+            _worldPacket.WriteUInt32((uint)Reason);
         }
 
         public ObjectGuid VendorGUID;
